Update 0001-two-sum.py
Fork to suggest changes to the code to make it more optimized.
The code is shorter, cleaner, and uses pre-built libraries which is a good way to demonstrate knowledge of multiple modules. pre-built module is also optimized for performance so, it will make the code quicker than it is before.